♻️ Refactor: 서버, 클라이언트 컴포넌트 분리 - 모바일, 웹 카테고리 - 모바일, 웹 최근 본 상품 페이지
