+++ src/app/categoryPage/web/page.tsx
@@ -1,0 +1,5 @@
+import CategoryPage from '@/app/categoryPage/web/categoryPage';
+
+export default function Page() {
+  return <CategoryPage />;
+}
