+++ src/styles/categoryPage/mobile/categoryHeader.module.css
@@ -1,0 +1,66 @@
+@import url(../../reset/modern-reset.css);
+@import url(../../globals.css);
+
+.header {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-self: center;
+  padding-block: 21px;
+  padding-inline: 16px;
+}
+
+.header_left {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+  font-size: 12px;
+}
+
+.filter_icon {
+  width: 24px;
+  height: 24px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.filter_text {
+  font-size: 12px;
+  color: var(--color-text);
+}
+
+.header_center {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.sort_select,
+.count_select {
+  padding: 2px;
+  padding-block: 4px;
+  border: 1px solid #ddd;
+  border-radius: 4px;
+  font-size: 10px;
+  background: #fff;
+  max-width: 60px;
+  max-height: 24px;
+  font-weight: 500;
+  align-items: center;
+}
+
+.header_right {
+  display: flex;
+  align-items: center;
+  margin-left: 5px;
+}
+
+.view_toggle {
+  background: none;
+  border: none;
+  font-size: 18px;
+  cursor: pointer;
+  padding: 4px;
+}
