+++ src/styles/categoryPage/mobile/categoryPage.module.css
@@ -1,0 +1,20 @@
+@import url(../../reset/modern-reset.css);
+@import url(../../globals.css);
+
+.container {
+  max-width: 480px;
+  min-width: 375px;
+  margin: 0 auto;
+  background: #fcfcfc;
+  min-height: 100vh;
+
+  a {
+    text-decoration: none;
+    color: inherit;
+  }
+
+  :hover {
+    text-decoration: none;
+    color: inherit;
+  }
+}
