+++ src/styles/categoryPage/web/categoryPage.module.css
@@ -1,5 +1,6 @@
-@import url(../styles/reset/modern-reset.css);
-@import url(../styles/globals.css);
+@import url(../../reset/modern-reset.css);
+@import url(../../globals.css);
+
 .container {
   width: 100%;
 }
@@ -99,14 +100,14 @@
   font-weight: 600;
 }
 
-/* 간단한 상품 카드 */
-.simple_product_card {
+/*상품 카드 */
+.product_card {
   border-radius: 8px;
   flex: 0 0 180px;
   background: #fcfcfc;
 }
 
-.simple_product_image img {
+.product_image img {
   border: 1px solid #b6c5c5;
   border-radius: 8px;
   width: 100%;
@@ -115,20 +116,20 @@
   margin-bottom: 20px;
 }
 
-.simple_product_name {
+.product_name {
   font-size: 16px;
   color: var(--color-text);
   margin-bottom: 5px;
   font-weight: 600;
 }
 
-.simple_product_price {
+.product_price {
   font-size: 14px;
   margin-bottom: 8px;
   font-weight: 600;
 }
 
-.simple_product_actions {
+.product_actions {
   font-size: 10px;
   color: #6b7878;
 }
