+++ src/styles/recentPage/mobile/recentProduct.module.css
@@ -1,27 +1,7 @@
-@import url(../styles/reset/modern-reset.css);
-@import url(../styles/globals.css);
+@import url(../../reset/modern-reset.css);
+@import url(../../globals.css);
 
-.recent_products_wrapper {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  z-index: 1000;
-}
-
-.recent_products_wrapper::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: -1;
-  display: none;
-}
-
+/* 컨테이너 */
 .container {
   width: 100%;
   max-width: 480px;
@@ -29,8 +9,6 @@
   margin: 0 auto;
   background-color: #fcfcfc;
   min-height: 100vh;
-  position: relative;
-  z-index: 1;
 }
 
 /* 헤더 */
@@ -56,7 +34,7 @@
   margin: 0;
 }
 
-.close_button {
+.closeButton {
   background: none;
   border: none;
   font-size: 24px;
@@ -68,7 +46,7 @@
   padding: 0;
 }
 
-.close_button:hover {
+.closeButton:hover {
   color: #333;
 }
 
@@ -78,7 +56,7 @@
 }
 
 /* 빈 상태 */
-.empty_state {
+.emptyState {
   text-align: center;
   padding: 80px 20px;
   color: var(--color-text);
@@ -87,11 +65,11 @@
 }
 
 /* 상품 리스트 */
-.product_list {
+.productList {
   padding: 0;
 }
 
-.product_item {
+.productItem {
   display: flex;
   align-items: flex-start;
   padding: 15px 10px;
@@ -103,19 +81,19 @@
   justify-content: center;
 }
 
-.product_item:nth-child(even) {
+.productItem:nth-child(even) {
   padding: 0px 10px;
   padding-right: 20px;
 }
 
-.product_item_border {
+.productItemBorder {
   display: flex;
   align-items: center;
   justify-content: center;
 }
 
 /* 상품 이미지 */
-.product_image {
+.productImage {
   flex-shrink: 0;
   width: 100px;
   height: 100px;
@@ -130,7 +108,7 @@
   padding: 4px;
 }
 
-.product_image_img {
+.productImageImg {
   width: 100%;
   height: 100%;
   object-fit: cover;
@@ -138,14 +116,14 @@
   background: #f5f5f5;
 }
 
-.image_placeholder {
+.imagePlaceholder {
   font-size: 12px;
   color: #6b7878;
   text-align: center;
 }
 
 /* 상품 정보 */
-.product_info {
+.productInfo {
   flex: 1;
   cursor: pointer;
   padding-block: 13px;
@@ -154,7 +132,7 @@
   overflow: hidden;
 }
 
-.brand_name {
+.brandName {
   font-size: 9px;
   color: #6b7878;
   margin-bottom: 4px;
@@ -165,7 +143,7 @@
   width: 100%;
 }
 
-.product_name {
+.productName {
   font-size: 14px;
   color: #333;
   font-weight: 500;
@@ -175,7 +153,7 @@
   text-overflow: ellipsis;
 }
 
-.product_price {
+.productPrice {
   font-size: 13px;
   color: var(--color-text);
   font-weight: 700;
@@ -183,7 +161,7 @@
 }
 
 /* 찜 버튼 */
-.like_button {
+.likeButton {
   background: none;
   border: none;
   display: flex;
@@ -196,14 +174,14 @@
   transition: color 0.2s ease;
 }
 
-.like_text {
+.likeText {
   font-size: 12px;
   color: #6b7878;
   font-weight: 700;
 }
 
 /* 삭제 버튼 */
-.remove_button {
+.removeButton {
   background: none;
   border: none;
   width: 16px;
@@ -215,40 +193,6 @@
   padding: 0;
 }
 
-.remove_button:hover {
+.removeButton:hover {
   color: var(--color-text);
-}
-
-/* 웹 스타일 (481px 이상) */
-@media (min-width: 481px) {
-  .recent_products_wrapper::before {
-    display: block;
-  }
-
-  .container {
-    position: fixed;
-    top: 0;
-    right: 0;
-    width: 400px;
-    max-width: 400px;
-    height: 100vh;
-    margin: 0;
-    border-radius: 0;
-    box-shadow: -4px 0 20px rgba(0, 0, 0, 0.15);
-    animation: slide_in_right 0.3s ease-out;
-  }
-
-  .content {
-    height: calc(100vh - 65px);
-    overflow-y: auto;
-  }
-}
-
-@keyframes slide_in_right {
-  from {
-    transform: translateX(100%);
-  }
-  to {
-    transform: translateX(0);
-  }
 }
